+++ docs/changes.md
@@ -1,4 +1,12 @@
 # 最新动态
+* 2019/08/24
+  * 更新文档。
+  * 完善mledit，支持readonly属性。
+
+* 2019/08/23
+  * 增加prop desc结构。
+  * 完善object/object default
+
 * 2019/08/22
   * 增加combo\_box\_ex控件(感谢大恒提供)
   * 增加属性WIDGET\_PROP\_FOCUSED表示当前的焦点状态，WIDGET\_PROP\_FOCUS属性为过时。